Extract various DataPort methods to package private utility.
